ローカルホストでのストリーミングOK
