+++ frontend_getyourtopic/assets/env.development
@@ -1,1 +1,2 @@
-API_SERVER=192.168.96.2:8000
+API_SERVER=127.0.0.1:8001
+# API_SERVER=192.168.96.2:8000
