+++ frontend_getyourtopic/lib/repository/get_topic_repository_test.dart
@@ -1,5 +1,6 @@
 import 'package:frontend_getyourtopic/model/get_topic.dart';
 import 'package:frontend_getyourtopic/repository/get_topic_repository.dart';
+import 'package:http/http.dart';
 
 class GetTopicRepositoryTest extends GetTopicRepository {
   @override
@@ -13,14 +14,7 @@ class GetTopicRepositoryTest extends GetTopicRepository {
   }
 
   @override
-  Stream<String> getTopicStream(String prompt, {String? pictureBase64}) async* {
-    // String responseString =
-    //     "Streaming from getTopicStream! Can you read this? prompt:$prompt / pictureBase64:$pictureBase64";
-
-    // for (var i = 0; i < responseString.length; i++) {
-    //   await Future.delayed(const Duration(milliseconds: 10));
-    //   yield responseString[i];
-    // }
+  Future<ByteStream> getTopicStream(String prompt, {String? pictureBase64}) {
     throw UnimplementedError();
   }
 }
